refactor: replace MagicPodRunner with direct implementation in GetBatchRuns
