+++ test/commands/get-batch-runs.test.ts
@@ -87,7 +87,7 @@ describe('get-batch-runs', () => {
     // Check stdout messesges
     expect(stdout).to.contain('INFO  [GcsStore]')
     expect(stdout).to.contain('INFO  [BigqueryExporter]')
-    expect(stdout).to.contain("INFO  [MagicPodRunner] Done execute 'magicpod'. status: success")
+    expect(stdout).to.contain("INFO  Done execute 'magicpod'. status: success")
 
     // Check output exists on GCS and BigQuery
     const [rows] = await bigquery.dataset('fake-dataset').table('test_report').getRows()
@@ -116,7 +116,7 @@ describe('get-batch-runs', () => {
 
     // Check stdout messesges
     expect(stdout).to.contain('INFO  [LocalStore]')
-    expect(stdout).to.contain("INFO  [MagicPodRunner] Done execute 'magicpod'. status: success")
+    expect(stdout).to.contain("INFO  Done execute 'magicpod'. status: success")
 
     // Check output exists on local
     const outputFiles = await fs.readdir('output')
@@ -150,11 +150,11 @@ describe('get-batch-runs', () => {
     ])
 
     // Check stdout messesges
-    expect(stdout).to.contain('INFO  [MagicPodRunner] --- Enable DEBUG mode ---')
+    expect(stdout).to.contain('INFO  --- Enable DEBUG mode ---')
     expect(stdout).to.contain('INFO  [NullStore] Detect DEBUG mode, nothing is used instead.')
     expect(stdout).to.contain('DEBUG [MagicPodClient]')
     expect(stdout).to.contain('INFO  [NullStore] Detect DEBUG mode, skip saving lastRun.')
-    expect(stdout).to.contain("INFO  [MagicPodRunner] Done execute 'magicpod'. status: success")
+    expect(stdout).to.contain("INFO  Done execute 'magicpod'. status: success")
 
     // Check output exists on local but lastRun does not exist
     const outputFiles = await fs.readdir(path.join('output'))
